+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
 ### Version 1.0.1:
+
 - Add ability to request otp.
 - update to the Msegat class structure.
+- add `log` config that allow you to send sms to log directly to use it you should add it to `env`
+  as `MSEGAT_LOG = true`.